requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.39.0
Django==5.0.2
django-ckeditor==6.7.1
django-cloudinary-storage==0.3.0
django-js-asset==2.2.0
gunicorn==21.2.0
idna==3.6
packaging==23.2
pillow==10.2.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
